celery_django/requirements.txt
Django==4.2.10
celery==5.3.6
django-celery-beat==2.5.0
redis==5.0.2
djangorestframework==3.14.0
drf-spectacular==0.27.1
gunicorn==21.2.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
backoff==2.2.1
requests==2.31.0
drf-spectacular==0.27.1
